+++ relayer-cli/src/commands/start_multi.rs
@@ -1,11 +1,11 @@
-use abscissa_core::{Command, Options, Runnable};
+use abscissa_core::{config, Command, Options, Runnable};
 
 use ibc::ics24_host::identifier::ChainId;
-use ibc_relayer::{config::Config, supervisor::Supervisor};
+use ibc_relayer::supervisor::Supervisor;
 
-use crate::commands::cli_utils::ChainHandlePair;
 use crate::conclude::Output;
 use crate::prelude::*;
+use crate::{application::CliApp, commands::cli_utils::ChainHandlePair};
 
 #[derive(Clone, Command, Debug, Options)]
 pub struct StartMultiCmd {
@@ -18,17 +18,21 @@ pub struct StartMultiCmd {
 
 impl Runnable for StartMultiCmd {
     fn run(&self) {
-        let config = (*app_config()).clone();
+        let config = app_config();
 
-        match start_multi(config, &self.src_chain_id, &self.dst_chain_id) {
+        match start_multi(&config, &self.src_chain_id, &self.dst_chain_id) {
             Ok(output) => output.exit(),
             Err(e) => Output::error(format!("{}", e)).exit(),
         }
     }
 }
 
-fn start_multi(config: Config, chain_a: &ChainId, chain_b: &ChainId) -> Result<Output, BoxError> {
-    let chains = ChainHandlePair::spawn(&config, chain_a, chain_b)?;
+fn start_multi(
+    config: &config::Reader<CliApp>,
+    chain_a: &ChainId,
+    chain_b: &ChainId,
+) -> Result<Output, BoxError> {
+    let chains = ChainHandlePair::spawn(config, chain_a, chain_b)?;
     let supervisor = Supervisor::spawn(chains.src, chains.dst)?;
     supervisor.run()?;